+++ src/main/java/com/example/sights/service/SightService.java
@@ -9,9 +9,9 @@ public interface SightService {
 
     List<Sight> readAll();
 
-    Sight read(int id);
+    Sight read(Long id);
 
-    boolean update(Sight sight, int id);
+    boolean update(Sight sight, Long id);
 
-    boolean delete(int id);
+    boolean delete(Long id);
 }
